[llvm] Improve system library flag detection.
